riscv: Move cpufeature.h macros into their own header
mainline inclusion
from mainline-v6.13
commit 010e12aa4925b36700ebacb763a7e6cfd771d9a2
category: feature
bugzilla: https://github.com/RVCK-Project/rvck-olk/issues/36

--------------------------------

asm/cmpxchg.h will soon need riscv_has_extension_unlikely() macros and
then needs to include asm/cpufeature.h which introduces a lot of header
circular dependencies.

So move the riscv_has_extension_XXX() macros into their own header which
prevents such circular dependencies by including a restricted number of
headers.

Link: https://lore.kernel.org/r/20241103145153.105097-2-alexghiti@rivosinc.com
